+++ packages/firestore/test/util/equality_matcher.ts
@@ -18,8 +18,8 @@ import { use } from 'chai';
 
 /**
  * @file This file provides a helper function to add a matcher that matches
- * based on an objects equals method.  If the equals method is present one
- * either object it is used to determine equality, else mocha's default equals
+ * based on an objects isEqual method.  If the isEqual method is present one
+ * either object it is used to determine equality, else mocha's default isEqual
  * implementation is used.
  */
 
@@ -27,11 +27,7 @@ function customDeepEqual(left, right) {
   /**
    * START: Custom compare logic
    */
-  // Internally we use .equals(), but our public API uses .isEqual() so
-  // we handle both.
-  if (left && typeof left.equals === 'function') return left.equals(right);
   if (left && typeof left.isEqual === 'function') return left.isEqual(right);
-  if (right && typeof right.equals === 'function') return right.equals(left);
   if (right && typeof right.isEqual === 'function') return right.isEqual(left);
   /**
    * END: Custom compare logic
@@ -42,14 +38,15 @@ function customDeepEqual(left, right) {
     typeof right === 'number' &&
     isNaN(left) &&
     isNaN(right)
-  )
+  ) {
     return true;
-  if (typeof left != typeof right) return false; // needed for structurally different objects
+  }
+  if (typeof left !== typeof right) return false; // needed for structurally different objects
   if (Object(left) !== left) return false; // primitive values
-  var keys = Object.keys(left);
-  if (keys.length != Object.keys(right).length) return false;
-  for (var i = 0; i < keys.length; i++) {
-    var key = keys[i];
+  const keys = Object.keys(left);
+  if (keys.length !== Object.keys(right).length) return false;
+  for (let i = 0; i < keys.length; i++) {
+    const key = keys[i];
     if (!Object.prototype.hasOwnProperty.call(right, key)) return false;
     if (!customDeepEqual(left[key], right[key])) return false;
   }
@@ -60,7 +57,7 @@ export function addEqualityMatcher() {
   let originalFunction;
   beforeEach(() => {
     use((chai, utils) => {
-      var Assertion = chai.Assertion;
+      const Assertion = chai.Assertion;
 
       const assertEql = _super => {
         originalFunction = originalFunction || _super;